8 frontend tests passing 33% coverage
