+++ frontend/src/__tests__/home-tests/Chogan.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
-import HomeListView from '../components/home/HomeListView';
-import Chogan from '../Chogan';
+import HomeListView from '../../components/home/HomeListView';
+import Chogan from '../../Chogan';
 
 describe('Chogan Tests', () => {
