+++ frontend/src/__tests__/home-tests/HomeListView.test.tsx
@@ -1,17 +1,23 @@
 import * as React from 'react';
 import { mount} from 'enzyme';
-import ProjectList from '../components/home/ProjectList.component';
+import HomeListView from '../../components/home/HomeListView';
+import { store } from '../../redux/store';
+import { Provider } from 'react-redux';
 
-describe('ProjectList Tests', () => {
+describe('HomeListView tests', () => {
     let wrapper:any;
     let projs:any;
     beforeEach(() => {
         projs = [{projectName: 'one', projectDescription: "one"}, {projectName: 'two', projectDescription: "two"}];
-        wrapper = mount(<ProjectList projects={projs} />);
+        wrapper = mount(
+          <Provider store={store} > 
+            <HomeListView projects={projs} /> 
+          </Provider>
+        );
      });
     it('shallow rendering project-name', () => {
         
-        const actual = wrapper.find('.project-name');
+        const actual = wrapper.find('.project-list-item');
         expect(actual.length).toBeGreaterThanOrEqual(projs.length);
     });
   });
